feat!: split the transport seam — stdio-only images shed netty (TJC-2114)
TransportBackend keeps serveStdio + randomId; serveHttp moves to the new
HttpTransportBackend trait. runHttp() takes it as a using parameter and
the TransportRunner[Http] given is conditional on it, so a stdio-only
program has no reachable call path into the HTTP stack — GraalVM
native-image reachability and Scala.js DCE drop zio-http/netty entirely.

- JVM: JvmTransportBackend is now stdio-only and netty-free; the HTTP
  implementation moves verbatim to JvmHttpBackend (given httpInstance),
  which also gains deliberate netty channel-type selection: AUTO on a
  normal JVM (unchanged), NIO under native-image (AUTO's runtime
  epoll/kqueue/io_uring probing breaks closed-world analysis), with a
  -Dfastmcp.http.channelType override. Server.customized with
  NettyConfig.default is behavior-identical to Server.defaultWith.
- JS: JsTransportBackend provides both givens; ExportsJs unchanged.
- Source-compatible for import com.tjclp.fastmcp.{*, given} users;
  binary-breaking (CHANGELOG note) — last window at 1.0.0-RC2.

Gates: jvm+js test suites, checkFormat, and the official conformance
suite (73/73 active, both platforms, baselines unchanged) all green.

Part of GH #66.
